rename for multiple watchfaceservices in project
