+++ apps/wear/watchfaces/SmartDriveWatchFace/app/.gitignore
