+++ apps/wear/watchfaces/SmartDriveWatchFace/app/proguard-rules.pro
