+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.permobil.smartdrivemx2digital">
+    package="com.permobil.smartdrive.watchface">
 
     <uses-feature android:name="android.hardware.type.watch" />
 
@@ -32,11 +32,11 @@
             android:value="true" />
 
         <activity
-            android:name=".WatchfaceLayoutActivity"
-            android:label="@string/title_activity_watchface_layout" />
+            android:name=".DigitalWatchFaceLayoutActivity"
+            android:label="@string/digital_activity_watchface_layout" />
 
         <service
-            android:name=".ComplicationWatchFaceService"
+            android:name=".DigitalWatchFaceService"
             android:label="@string/my_analog_name"
             android:permission="android.permission.BIND_WALLPAPER">
             <meta-data
@@ -61,7 +61,7 @@
 
         <activity android:name="android.support.wearable.complications.ComplicationHelperActivity" />
         <activity
-            android:name=".ComplicationConfigActivity"
+            android:name=".DigitalWatchFaceConfigActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="com.example.android.wearable.watchface.CONFIG_COMPLICATION" />
