+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/drawable-hdpi/ic_battery_full.png
