+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/drawable-mdpi/ic_battery_full.png
