+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/drawable-mdpi/smartdrive_button.png
