+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/drawable-nodpi/added_complication.png
