+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/drawable-nodpi/preview_watch_face.png
