+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/drawable-xhdpi/ic_battery_full.png
