+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/drawable-xxhdpi/ic_battery_full.png
