+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/drawable-xxxhdpi/ic_battery_full.png
