+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/drawable/black_border_background.xml
