+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/drawable/settings_watch_face_preview_arms_and_ticks.xml
