+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/layout/activity_config.xml
