+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/mipmap-hdpi/ic_launcher.png
