+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/mipmap-mdpi/ic_launcher.png
