+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/mipmap-xhdpi/ic_launcher.png
