+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/mipmap-xxhdpi/ic_launcher.png
