+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/values-round/strings.xml
