+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/values/color.xml
