+++ apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/resources/sentry.properties
