+++ apps/wear/watchfaces/SmartDriveWatchFace/gradle.properties
