+++ apps/wear/watchfaces/SmartDriveWatchFace/local.properties
