+++ apps/wear/watchfaces/SmartDriveWatchFace/settings.gradle
